requirements.txt
# PCB Defect Detection - Python Dependencies

# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0

# Web Framework
flask>=2.3.0
werkzeug>=2.3.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0

# Data Science
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
tqdm>=4.65.0

# Dataset Download
kagglehub>=0.1.0

# Production Server (optional)
gunicorn>=21.0.0

# Development Tools (optional)
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0
